move client to api/v1/
