+++ command/find.go
@@ -9,8 +9,8 @@ import (
 	"github.com/libp2p/go-libp2p-core/host"
 	"github.com/urfave/cli/v2"
 
-	httpclient "github.com/filecoin-project/storetheindex/client/http"
-	p2pclient "github.com/filecoin-project/storetheindex/client/libp2p"
+	httpclient "github.com/filecoin-project/storetheindex/api/v1/client/http"
+	p2pclient "github.com/filecoin-project/storetheindex/api/v1/client/libp2p"
 	"github.com/filecoin-project/storetheindex/internal/finder"
 	"github.com/filecoin-project/storetheindex/server/net"
 )
